policy: Split API into individual files
