+++ pkg/policy/api/cidr.go
@@ -1,0 +1,44 @@
+// Copyright 2016-2017 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api
+
+// CIDR specifies a block of IP addresses.
+// Example: 192.0.2.1/32
+type CIDR string
+
+// CIDRMatchAll is a []CIDR that matches everything
+var CIDRMatchAll = []CIDR{CIDR("0.0.0.0/0"), CIDR("::/0")}
+
+// CIDRRule is a rule that specifies a CIDR prefix to/from which outside
+// communication  is allowed, along with an optional list of subnets within that
+// CIDR prefix to/from which outside communication is not allowed.
+type CIDRRule struct {
+	// CIDR is a CIDR prefix / IP Block.
+	//
+	Cidr CIDR `json:"cidr"`
+
+	// ExceptCIDRs is a list of IP blocks which the endpoint subject to the rule
+	// is not allowed to initiate connections to. These CIDR prefixes should be
+	// contained within Cidr. These exceptions are only applied to the Cidr in
+	// this CIDRRule, and do not apply to any other CIDR prefixes in any other
+	// CIDRRules.
+	//
+	// +optional
+	ExceptCIDRs []CIDR `json:"except,omitempty"`
+
+	// Generated indicates whether the rule was generated based on other rules
+	// or provided by user
+	Generated bool `json:"-"`
+}
