+++ pkg/policy/api/service.go
@@ -1,0 +1,38 @@
+// Copyright 2016-2017 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api
+
+// ServiceSelector is a label selector for k8s services
+type ServiceSelector EndpointSelector
+
+// Service wraps around selectors for services
+type Service struct {
+	// K8sServiceSelector selects services by k8s labels and namespace
+	K8sServiceSelector *K8sServiceSelectorNamespace `json:"k8sServiceSelector,omitempty"`
+	// K8sService selects service by name and namespace pair
+	K8sService *K8sServiceNamespace `json:"k8sService,omitempty"`
+}
+
+// K8sServiceNamespace is an abstraction for the k8s service + namespace types.
+type K8sServiceNamespace struct {
+	ServiceName string `json:"serviceName,omitempty"`
+	Namespace   string `json:"namespace,omitempty"`
+}
+
+// K8sServiceSelectorNamespace wraps service selector with namespace
+type K8sServiceSelectorNamespace struct {
+	Selector  ServiceSelector `json:"selector"`
+	Namespace string          `json:"namespace,omitempty"`
+}
